ms/boardManagementService/src/test/java/com/elysium/reddot/ms/board/infrastructure/inbound/rest/processor/DeleteBoardProcessorTest.java
package com.elysium.reddot.ms.board.infrastructure.inbound.rest.processor;

import com.elysium.reddot.ms.board.application.data.dto.ApiResponseDTO;
import com.elysium.reddot.ms.board.application.data.dto.BoardDTO;
import com.elysium.reddot.ms.board.application.service.BoardApplicationServiceImpl;
import org.apache.camel.CamelContext;
import org.apache.camel.Exchange;
import org.apache.camel.impl.DefaultCamelContext;
import org.apache.camel.support.DefaultExchange;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
class DeleteBoardProcessorTest {

    private DeleteBoardProcessor deleteBoardProcessor;

    @Mock
    private BoardApplicationServiceImpl boardService;

    private CamelContext camelContext;

    @BeforeEach
    void setUp() {
        camelContext = new DefaultCamelContext();
        deleteBoardProcessor = new DeleteBoardProcessor(boardService);
    }

    @Test
    @DisplayName("given valid boardId when deleteBoard is called then board deleted Successfully")
    void givenValidBoardId_whenDeleteBoard_thenBoardIsDeletedSuccessfully() {
        // arrange
        Long boardId = 1L;
        BoardDTO boardDTO = new BoardDTO(boardId, "name", "Name", "Board description");

        ApiResponseDTO expectedApiResponse = new ApiResponseDTO(HttpStatus.OK.value(),
                "Board with id " + boardId + " deleted successfully", boardDTO);

        Exchange exchange = new DefaultExchange(camelContext);
        exchange.getIn().setHeader("CamelHttpUri", "/boards/" + boardId);
        exchange.getIn().setHeader("id", boardId);

        // mock
        when(boardService.getBoardById(boardId)).thenReturn(boardDTO);
        doNothing().when(boardService).deleteBoardById(boardId);

        // act
        deleteBoardProcessor.process(exchange);
        ApiResponseDTO actualApiResponse = exchange.getMessage().getBody(ApiResponseDTO.class);

        // assert
        assertEquals(expectedApiResponse.getStatus(), actualApiResponse.getStatus());
        assertEquals(expectedApiResponse.getMessage(), actualApiResponse.getMessage());
        assertEquals(expectedApiResponse.getData(), actualApiResponse.getData());

        // verify
        verify(boardService).deleteBoardById(boardId);
    }

}
